Move metadata classes to meta package
