+++ servers/src/main/java/tachyon/master/next/filesystem/FileSystemMaster.java
@@ -37,6 +37,12 @@ import tachyon.master.next.block.BlockMaster;
 import tachyon.master.next.block.meta.BlockWorkerInfo;
 import tachyon.master.next.block.meta.UserBlockInfo;
 import tachyon.master.next.block.meta.UserBlockLocation;
+import tachyon.master.next.filesystem.meta.DependencyMap;
+import tachyon.master.next.filesystem.meta.FileBlockInfo;
+import tachyon.master.next.filesystem.meta.Inode;
+import tachyon.master.next.filesystem.meta.InodeDirectory;
+import tachyon.master.next.filesystem.meta.InodeFile;
+import tachyon.master.next.filesystem.meta.InodeTree;
 import tachyon.thrift.BlockInfoException;
 import tachyon.thrift.ClientBlockInfo;
 import tachyon.thrift.ClientDependencyInfo;
