+++ servers/src/main/java/tachyon/master/next/filesystem/meta/Inode.java
@@ -13,7 +13,7 @@
  * the License.
  */
 
-package tachyon.master.next.filesystem;
+package tachyon.master.next.filesystem.meta;
 
 import tachyon.thrift.ClientFileInfo;
