+++ servers/src/main/java/tachyon/master/next/filesystem/meta/InodeTree.java
@@ -13,7 +13,7 @@
  * the License.
  */
 
-package tachyon.master.next.filesystem;
+package tachyon.master.next.filesystem.meta;
 
 import java.util.ArrayList;
 import java.util.HashMap;
